requirements.txt
fastapi>=0.115.0
jinja2>=3.1.4
openai>=1.40.0
playwright>=1.49.0
pybars3>=0.9.7
pypdf>=5.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.18
uvicorn[standard]>=0.32.0
